recode/sandbox-1.0.0.145/examples/oss-upload/pyproject.toml
[project]
name = "oss-upload"
version = "0.1.0"
description = "OSS upload example for agent-sandbox (AWS S3, Volcengine TOS)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agent-sandbox",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
    "tos>=2.6.0",
]

[tool.uv.sources]
agent-sandbox = { path = "../../sdk/python", editable = true }
